Update and rename PlayingScene.h to PlayScene.h
