--- src/PlayingScene.h
@@ -1,20 +1,0 @@
-#ifndef __PLAYING_SCENE_H__
-#define __PLAYING_SCENE_H__
-
-#include "cocos2d.h"
-USING_NS_CC;
-
-class JoyStick;
-
-class PlayingScene : public cocos2d::Layer
-{
-public:
-    static cocos2d::Scene* createScene();
-	void PlayingScene::addTouchEvent();
-	virtual bool init();
-	CCPoint convertToGL(CCSet *pTouches);
-    CREATE_FUNC(PlayingScene);
-   
-};
-
-#endif // __HELLOWORLD_SCENE_H__
